move public ApodiniError initializer to ApodiniExtension
